Refactor devices.js
